Api/requirements.txt
dynaconf
flask
flask-restful
flask-sqlalchemy
gunicorn
mysql-connector-python